+++ lib/ui/widgets/QrCodePanel.dart
@@ -9,7 +9,9 @@ import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:intl/intl.dart';
 import 'package:rokwire_plugin/model/event2.dart';
+import 'package:rokwire_plugin/model/group.dart';
 import 'package:rokwire_plugin/service/events2.dart';
+import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
@@ -17,7 +19,7 @@ import 'package:rokwire_plugin/utils/image_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:share/share.dart';
 
-class Event2QrCodePanel extends StatefulWidget { //TBD localize
+class QrCodePanel extends StatefulWidget { //TBD localize
   //final Event2? event;
   //const Event2QrCodePanel({required this.event});
 
@@ -30,7 +32,7 @@ class Event2QrCodePanel extends StatefulWidget { //TBD localize
   final String? title;
   final String? description;
 
-  const Event2QrCodePanel({Key? key,
+  const QrCodePanel({Key? key,
     required this.deepLinkUrl,
 
     required this.saveFileName,
@@ -41,31 +43,41 @@ class Event2QrCodePanel extends StatefulWidget { //TBD localize
     this.description,
   });
 
-  factory Event2QrCodePanel.fromEvent(Event2? event, {Key? key}) => Event2QrCodePanel(
+  factory QrCodePanel.fromEvent(Event2? event, {Key? key}) => QrCodePanel(
     key: key,
     deepLinkUrl: Events2.eventDetailUrl(event),
       saveFileName: 'event - ${event?.name}',
       saveWatermarkText: event?.name,
       saveWatermarkStyle: TextStyle(fontFamily: Styles().fontFamilies.bold, fontSize: 64, color: Styles().colors.textSurface),
-    title: Localization().getStringEx('panel.event_qr_code.event.title', 'Share this event'),
-    description: Localization().getStringEx('panel.event_qr_code.event.description', 'Invite others to view this event by sharing a link or the QR code after saving it to your photo library.'),
+    title: Localization().getStringEx('panel.qr_code.event.title', 'Share this event'),
+    description: Localization().getStringEx('panel.qr_code.event.description', 'Invite others to view this event by sharing a link or the QR code after saving it to your photo library.'),
   );
 
-  factory Event2QrCodePanel.fromFilterParam(Event2FilterParam filterParam, {Key? key}) => Event2QrCodePanel(
+  factory QrCodePanel.fromEventFilterParam(Event2FilterParam filterParam, {Key? key}) => QrCodePanel(
     key: key,
     deepLinkUrl: Events2.eventsQueryUrl(filterParam.toUriParams()),
       saveFileName: "events ${DateFormat('yyyy-MM-dd HH.mm.ss').format(DateTime.now())}",
       saveWatermarkText: filterParam.buildDescription().map((span) => span.toPlainText()).join(),
       saveWatermarkStyle: TextStyle(fontFamily: Styles().fontFamilies.regular, fontSize: 32, color: Styles().colors.textSurface),
-    title: Localization().getStringEx('panel.event_qr_code.query.title', 'Share this event set'),
-    description: Localization().getStringEx('panel.event_qr_code.query.description', 'Invite others to view this set of filtered events by sharing a link or the QR code after saving it to your photo library.'),
+    title: Localization().getStringEx('panel.qr_code.event_query.title', 'Share this event set'),
+    description: Localization().getStringEx('panel.qr_code.event_query.description', 'Invite others to view this set of filtered events by sharing a link or the QR code after saving it to your photo library.'),
+  );
+
+  factory QrCodePanel.fromGroup(Group? group, {Key? key}) => QrCodePanel(
+    key: key,
+    deepLinkUrl: '${Groups().groupDetailUrl}?group_id=${group?.id}',
+      saveFileName: 'group - ${group?.title}',
+      saveWatermarkText: group?.title,
+      saveWatermarkStyle: TextStyle(fontFamily: Styles().fontFamilies.bold, fontSize: 64, color: Styles().colors.textSurface),
+    title: Localization().getStringEx('panel.qr_code.group.title', 'Share this group'),
+    description: Localization().getStringEx('panel.qr_code.group.description.label', 'Invite others to join this group by sharing a link or the QR code after saving it to your photo library.'),
   );
 
   @override
-  _EventQrCodePanelState createState() => _EventQrCodePanelState();
+  State<StatefulWidget> createState() => _QrCodePanelState();
 }
 
-class _EventQrCodePanelState extends State<Event2QrCodePanel> {
+class _QrCodePanelState extends State<QrCodePanel> {
   static final int _imageSize = 1024;
   Uint8List? _qrCodeBytes;
 
@@ -92,7 +104,7 @@ class _EventQrCodePanelState extends State<Event2QrCodePanel> {
     Analytics().logSelect(target: "Save Event QR Code");
 
     if (_qrCodeBytes == null) {
-      AppAlert.showDialogResult(context, Localization().getStringEx("panel.event_qr_code.alert.no_qr_code.msg", "There is no QR Code"));
+      AppAlert.showDialogResult(context, Localization().getStringEx("panel.qr_code.alert.no_qr_code.msg", "There is no QR Code"));
     } else {
       Uint8List? updatedImageBytes = await ImageUtils.applyLabelOverImage(_qrCodeBytes, widget.saveWatermarkText,
         width: _imageSize.toDouble(),
@@ -106,11 +118,11 @@ class _EventQrCodePanelState extends State<Event2QrCodePanel> {
 
       const String destinationMacro = '{{Destination}}';
       String messageSource = (result
-          ? (Localization().getStringEx("panel.event_qr_code.alert.save.success.msg", "Successfully saved qr code in $destinationMacro"))
-          : Localization().getStringEx("panel.event_qr_code.alert.save.fail.msg", "Failed to save qr code in $destinationMacro"));
+          ? (Localization().getStringEx("panel.qr_code.alert.save.success.msg", "Successfully saved qr code in $destinationMacro"))
+          : Localization().getStringEx("panel.qr_code.alert.save.fail.msg", "Failed to save qr code in $destinationMacro"));
       String destinationTargetText = (defaultTargetPlatform == TargetPlatform.android)
-          ? Localization().getStringEx("panel.event_qr_code.alert.save.success.pictures", "Pictures")
-          : Localization().getStringEx("panel.event_qr_code.alert.save.success.gallery", "Gallery");
+          ? Localization().getStringEx("panel.qr_code.alert.save.success.pictures", "Pictures")
+          : Localization().getStringEx("panel.qr_code.alert.save.success.gallery", "Gallery");
       String message = messageSource.replaceAll(destinationMacro, destinationTargetText);
       AppAlert.showDialogResult(context, message).then((value) {
         if(result) {
@@ -143,14 +155,14 @@ class _EventQrCodePanelState extends State<Event2QrCodePanel> {
                   padding: EdgeInsets.only(top: 24),
                   child: ((_qrCodeBytes != null)
                       ? Semantics(
-                    label: Localization().getStringEx('panel.event_qr_code.code.hint', "QR code image"),
+                    label: Localization().getStringEx('panel.qr_code.code.hint', "QR code image"),
                     child: Container(
                       decoration: BoxDecoration(color: Styles().colors.white, borderRadius: BorderRadius.all(Radius.circular(5))),
                       padding: EdgeInsets.all(5),
                       child: Image.memory(
                         _qrCodeBytes!,
                         fit: BoxFit.fitWidth,
-                        semanticLabel: Localization().getStringEx("panel.event_qr_code.primary.heading.title", "Promotion Key"),
+                        semanticLabel: Localization().getStringEx("panel.qr_code.primary.heading.title", "Promotion Key"),
                       ),
                     ),
                   )
@@ -169,7 +181,7 @@ class _EventQrCodePanelState extends State<Event2QrCodePanel> {
                 Padding(
                   padding: EdgeInsets.only(top: 24, bottom: 12),
                   child: RoundedButton(
-                    label: Localization().getStringEx('panel.event_qr_code.button.save.title', 'Save QR Code'),
+                    label: Localization().getStringEx('panel.qr_code.button.save.title', 'Save QR Code'),
                     hint: '',
                     textStyle: Styles().textStyles.getTextStyle("widget.title.regular.fat"),
                     backgroundColor: Styles().colors.background,
@@ -180,14 +192,12 @@ class _EventQrCodePanelState extends State<Event2QrCodePanel> {
                 Padding(
                   padding: EdgeInsets.only(bottom: 12),
                   child: RoundedButton(
-                    label: Localization().getStringEx('panel.event_qr_code.button.share.title', 'Share Link'),
+                    label: Localization().getStringEx('panel.qr_code.button.share.title', 'Share Link'),
                     hint: '',
                     textStyle: Styles().textStyles.getTextStyle("widget.button.title.medium.fat"),
                     backgroundColor: Styles().colors.background,
                     borderColor: Styles().colors.fillColorSecondary,
                     onTap: _onTapShare,
-                    rightIcon: Styles().images.getImage('share-dark', excludeFromSemantics: true),
-                    rightIconPadding: EdgeInsets.only(right: 75),
                   ),
                 )
               ],